initial

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -1,10 +1,11 @@
 import { validateRequest } from "@/auth";
 import prisma from "@/lib/prisma";
-import { FollowerInfo } from "@/lib/types";
+import { LikeInfo } from "@/lib/types";
+import { error } from "console";
 
 export async function GET(
   req: Request,
-  { params: { userId } }: { params: { userId: string } },
+  { params: { postId } }: { params: { postId: string } },
 ) {
   try {
     const { user: loggedInUser } = await validateRequest();
@@ -13,32 +14,32 @@ export async function GET(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    const user = await prisma.user.findUnique({
-      where: { id: userId },
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
       select: {
-        followers: {
+        likes: {
           where: {
-            followerId: loggedInUser.id,
+            userId: loggedInUser.id,
           },
           select: {
-            followerId: true,
+            userId: true,
           },
         },
         _count: {
           select: {
-            followers: true,
+            likes: true,
           },
         },
       },
     });
 
-    if (!user) {
-      return Response.json({ error: "User not found" }, { status: 404 });
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
     }
 
-    const data: FollowerInfo = {
-      followers: user._count.followers,
-      isFollowedByUser: !!user.followers.length,
+    const data: LikeInfo = {
+      likes: post._count.likes,
+      isLikedByUser: !!post.likes.length,
     };
 
     return Response.json(data);
@@ -50,7 +51,7 @@ export async function GET(
 
 export async function POST(
   req: Request,
-  { params: { userId } }: { params: { userId: string } },
+  { params: { postId } }: { params: { postId: string } },
 ) {
   try {
     const { user: loggedInUser } = await validateRequest();
@@ -59,27 +60,43 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
     await prisma.$transaction([
-      prisma.follow.upsert({
+      prisma.like.upsert({
         where: {
-          followerId_followingId: {
-            followerId: loggedInUser.id,
-            followingId: userId,
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
           },
         },
         create: {
-          followerId: loggedInUser.id,
-          followingId: userId,
+          userId: loggedInUser.id,
+          postId,
         },
         update: {},
       }),
-      prisma.notification.create({
-        data: {
-          issuerId: loggedInUser.id,
-          recipientId: userId,
-          type: "FOLLOW",
-        },
-      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
     ]);
 
     return new Response();
@@ -91,7 +108,7 @@ export async function POST(
 
 export async function DELETE(
   req: Request,
-  { params: { userId } }: { params: { userId: string } },
+  { params: { postId } }: { params: { postId: string } },
 ) {
   try {
     const { user: loggedInUser } = await validateRequest();
@@ -100,18 +117,30 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
     await prisma.$transaction([
-      prisma.follow.deleteMany({
+      prisma.like.deleteMany({
         where: {
-          followerId: loggedInUser.id,
-          followingId: userId,
+          userId: loggedInUser.id,
+          postId,
         },
       }),
       prisma.notification.deleteMany({
         where: {
           issuerId: loggedInUser.id,
-          recipientId: userId,
-          type: "FOLLOW",
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
         },
       }),
     ]);
